Make use of action outputs and report on deploy failure reason
As things stand, if a deployment fails we also fail the monitoring action
which conceptually is not ideal.
In this PR we ensure the monitoring action completes successfully (unless
it truly fails) and it also attempts to capture a more specific reason for
a deploy failure and report it via a GH action output.

Part of https://github.com/fac/platform-issues/issues/2054