Better slack reporting when pre-flight checks fail
At the moment if preflight checks fail (for example a forward deploy check)
this is not clear in slack.
With this PR we make sure that the specific failure is reported back from this action
which then is captured at the calling GH action and reported back to slack.

See this slack thread demonstrating the problem we are trying to address with this PR
https://freeagent.slack.com/archives/C0DT7DGS1/p1676385750719519