Complete refactor
This is a complete refactor while maintaining the current behaviour.

We move all the logic into a number of classes:

`Deployment::Deployment`

This is responsible for checking the status of a state machine
execution. If it has failed for some reason then we find the fail event
and then pass the event onto the correct class for further parsing.
If no class exists then we return a generic error message.

`Deployment::Events::FailStateEntered`

This handles the case where we hit the Fail state.
It then extracts the error attempts to determine if it was sidekiq that
failed or not returning the appropriate message.

`Deployment::Events::LambdaFunctionFailed`

This handles the case where a lambda error causes the state machine
execution to terminate but doesn't actually reach the fail state.
Currently the only time this happens is during the Pre Flight Checks.

This happens to be by accident since looking at the definition then we
are supposed to go to the fail state but we are catching the error
incorrectly.