Add some basic rspec tests
This script is in need of refactoring as it is hard to follow the logic
making it difficult to make changes to.

This adds some very basic tests to test the following so we can begin
refactoring:

- Test when deploy successful
- Test failure due to Sidekiq not starting
- Test failure due to Forward deploy check failure

All other errors will result in a fairly generic error message and the
above should be enough to enable us to test the refactor.

The aim is once we have refactored we can add a better test suite.